rename Switch to Toggle
